feat:添加4个路由
