+++ index.html
@@ -11,6 +11,6 @@
 <body>
   <div id="root">测试用</div>
 </body>
-<script type="module" src='./main.tsx'></script>
+<script type="module" src='./src/main.tsx'></script>
 
 </html>
