--- src/router/RedirectTowelcome1.tsx
@@ -1,12 +1,0 @@
-import React, { useEffect } from 'react'
-
-import { useNavigate } from 'react-router-dom'
-
-export const RedirectTowelcome1 = () => {
-  const navigate = useNavigate()
-  useEffect(() => {
-    navigate('/')
-  }, [])
-
-  return <div>123211</div>
-}
